# api/auth/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/strike": {
            "post": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Создает новую жалобу на пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "strikes"
                ],
                "summary": "Создать жалобу",
                "parameters": [
                    {
                        "description": "Данные для создания жалобы",
                        "name": "strike",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.StrikeCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.StrikeResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Дублирующая жалоба",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/strike/range": {
            "get": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Возвращает список жалоб за указанный период. По умолчанию: from=начало времен, to=текущее время + 1 день, limit=20, offset=0",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "strikes"
                ],
                "summary": "Получить жалобы за период",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Начальная дата (RFC3339). По умолчанию: начало времен",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Конечная дата (RFC3339). По умолчанию: текущее время + 1 день",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Лимит",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Смещение",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.StrikesListResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/strike/type/{type}": {
            "get": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Возвращает список жалоб по типу нарушения",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "strikes"
                ],
                "summary": "Получить жалобы по типу",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Тип нарушения",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Лимит",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Смещение",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.StrikesListResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/strike/user/{user_id}": {
            "get": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Возвращает список жалоб на конкретного пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "strikes"
                ],
                "summary": "Получить жалобы на пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID пользователя",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Лимит",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Смещение",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.StrikesListResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/strike/user/{user_id}/stat": {
            "get": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Возвращает статистику по жалобам на пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "strikes"
                ],
                "summary": "Получить статистику жалоб пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID пользователя",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.StrikeStats"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/strike/{id}/status": {
            "put": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Обновляет статус жалобы (только для модераторов)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "strikes"
                ],
                "summary": "Обновить статус жалобы",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID жалобы",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления статуса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.StrikeStatusUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.StrikeResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Жалоба не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.StrikeCreateRequest": {
            "type": "object",
            "properties": {
                "reason": {
                    "type": "string"
                },
                "reporter_id": {
                    "type": "string"
                },
                "target_user_id": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/constants.StrikeType"
                }
            }
        },
        "dto.StrikeResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "moderator_id": {
                    "type": "string"
                },
                "moderator_note": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "reporter_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/constants.StrikeStatus"
                },
                "target_user_id": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/constants.StrikeType"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.StrikeStats": {
            "type": "object",
            "properties": {
                "last_strike_at": {
                    "type": "string"
                },
                "strike_types": {
                    "$ref": "#/definitions/dto.StrikeTypeStat"
                },
                "total_strikes": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.StrikesListResponse": {
            "type": "object",
            "properties": {
                "strikes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.StrikeResponse"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    }
}